pax and t5x builds now depend on jax instead of te and install te in their own build step
This is to address a dependency issue that just recently cropped up:

- T5x required seqio required TF-text required TF required
typing_extensions
- TE required pydantic required typing_extensions

When installed sequentially, typing_extensions is downgraded by t5x and
is in conflict with the one in TE. There exists a working combination of
versions, but it is only achievable if pip installed altogether. This
change allows deferring installation of a library so that it can be
collectively installed later.