Redesign job reporting system (#222)
The JAX unit test workflow is modified to serve as a model for future
unification of job reporting.

# TLDR

Each job produces its own
[sitrep](https://thepersimmongroup.com/situation-report-sitrep-template/)
and badges as artifacts; badges from a workflow run are packaged as a
gist; sitreps are compiled into a cohesive workflow summary.

# Problems to solve

- Our current CI job reporting is fragmented, and it is hard to grasp
what went wrong in a particular workflow run without carefully reading
the step summaries, many of which are not readable.
- It is difficult to trace badge updates and answer questions such as
"Why is the badge still green while the nightly failed."
- The `_publish_badge.yaml` workflow serves two purposes: job
post-processing and (optionally) publishing a badge. In fact, it is more
frequently used to do post-processing rather than publishing. Hence the
inconsistency between its name and functionality can be misleading.

# Key changes

1. Each job generates a `sitrep.json` file with at least a `summary`
key. The file will be uploaded as an artifact.
2. Any job can optionally generate shields.io badge endpoint JSON files.
3. A helper bash function `.github/workflows/scripts/to_json.sh` can
assist in 1 & 2.
4. A new reusable workflow `.github/workflows/_upload_badge.yaml` will
upload badge files (currently based on file name pattern) from all jobs
within a workflow run as a **new** gist.
5. A new reusable workflow `.github/workflows/_copy_gist.yaml` for
selectively copying files from one gist (e.g., the job gist) to another
one (e.g., the landing page badges gist).
6. A new reusable workflow `.github/workflows/_summary.yaml` will
download sitreps from all jobs and concatenate the summaries into a
Markdown job summary.

# Opt-in guide

To opt-in to the new paradigm, refer to
[_test_jax.yaml](https://github.com/NVIDIA/JAX-Toolbox/blob/main/.github/workflows/_test_jax.yaml)
or follow the steps below:
1. **job-scope changes**: in each job that runs tests, add a job step
that looks like:
```
      - name: Process results and generate report
        if: success() || failure()
        shell: bash -x -e {0}
        run: |
          # bring in utility functions
          source .github/workflows/scripts/to_json.sh

          # POSTPROCESSING SCRIPTS...

          # generate sitrep
          to_json \
            summary \
            errors total_tests passed_tests failed_tests \
          > sitrep.json

          # optional: generate badge endpoint
          schemaVersion=1 \
          label="${badge_label}" \
          message="${badge_message}" \
          color="${badge_color}" \
          to_json schemaVersion label message color \
          > ${{ env.BADGE_FILENAME_FULL }}

      - name: Upload artifacts
        if: success() || failure()
        uses: actions/upload-artifact@v3
        with:
          name: ${{ inputs.ARTIFACT_NAME }}-${{ matrix.GPU_ARCH }}
          path: |
            sitrep.json
            ${{ env.BADGE_FILENAME_FULL }}
```
2. **workflow-scope changes**: for an entire workflow, add the following
jobs. You can optionally adjust if the jobs should run regardless of
whether the test job succeeded or not using the
[`if`](https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idif)
key.
```
  upload-badge:
    needs: <JOBS>
    uses: ./.github/workflows/_upload_badge.yaml
    secrets: inherit

  report:
    needs: <JOBS>
    uses: ./.github/workflows/_summary.yaml
    secrets: inherit

  publish-badge:
    needs: [upload-badge]
    if: <PUBLISHING-CONDITIONS>
    uses: ./.github/workflows/_copy_gist.yaml
    with:
      SRC_GIST_ID: ${{ needs.upload-badge.outputs.GIST_ID }}
      DST_GIST_ID: ${{ vars.BADGE_ENDPOINT_GIST_ID }}
      FILES: '.*badge.*.json'
    secrets: inherit
```
3. Remove usage of the old `.github/workflows/_publish_badge.yaml`
workflow.