Adds a triage job that triggers if pax MGMN or t5x MGMN tests fail (#208)
This is a large addition so the highlights are summarized below:
* Adds a new `_triage.yaml` workflow that gets called if T5x MGMN or Pax
MGMN tests fail. It relies on a `:latest-verified` tag as the last known
working image.
* The `latest-verified` tag is added to the vanilla t5x and pax images
if their respective tests pass
* Adds `Dockerfile.ff` which creates a new image from a `BASE_IMAGE`
where all (or some) of the dependencies are fast forwarded to the
versions in `BROKEN_IMAGE`. The image layer produced by this is small
since no recompilation happens. Thus, this dockerfile is currently only
suitable for fast-forwarding libraries with python source only (e.g.,
t5x, paxml, praxis).
* Additionally, a `ff-summary` script is added to help inspect the
source offline.
* To signal failure or success, a new job called `outcome` was added to
_test_t5x.yaml and _test_pax.yaml in order to detect if the overall test
was a success
* Adds some utility functions for workflows in this file:
`.github/workflows/util.sh`; which can be reused on the command line or
in other workflows. (Was helpful in debugging the triage workflow)
* Removes the endpoint artifact upload in _publish_badge.yaml since it
was causing collisions on multiple re-runs of the same workflow in a
matrix. The artifact isn't really needed anyway since we upload it as a
gist and it's printed in the logs.
* Files a github issue if there are failures:
    * For pax: https://github.com/NVIDIA/JAX-Toolbox/issues/215
    * For t5x: https://github.com/NVIDIA/JAX-Toolbox/issues/218