Remove 1 reusable workflow: copy_gist (#400)
The reusable workflow `_copy_gist.yaml` is only used by
`_finalyze.yaml`. To workaround the 20-reusable-workflow limit of GitHub
Actions, we can save one workload by inlining it. The reusable workflow
file is still kept but renamed with one extra prepending `_` in case it
will become useful in the future.

Co-authored-by: Yu-Hang "Maxin" Tang <Tang.Maxin@gmail.com>