Fixes issue where dated rosetta nightlies were not tagged if tests failed (#446)
The issue was that if tests failed, the `publish-final` step wasn't run
b/c a failure in any of the `needs:` jobs would implicitly cause
`publish-final` to be skipped. So the ternary operator to conditionally
add `latest` wasn't run and `nightly-YYYY-MM-DD` was skipped.

Simple remedy is to introduce two publish steps. The code is repeated,
but it is much clearer than adding complicated `if:` logic