Update manifest and pin newer numpy in containers (#444)
As of #405 the presubmit CI is based on package versions listed in the
`manifest.yaml` that is committed to the repository. This has not been
updated for ~1 month, so the presubmit CI is testing ~1 month old
versions of the ecosystem. This PR updates it using the commit from
https://github.com/NVIDIA/JAX-Toolbox/tree/znightly-2024-01-03-7395605285,
generated by the nightly CI run.

Because this bumps the JAX version by ~1 month, we have to include fixes
for deprecations. In particular replacing `jax.random.KeyArray` with
plain `jax.Array`
(https://github.com/nvjax-svc-0/t5x/commit/4d5ec2fe4a665142de514ad08544c54ab4be6133).

The deprecated name is used in older versions of the `chex` package,
which are being selected by pip's dependency resolver despite newer
versions being available. We avoid this by giving pip a helping hand and
nudging it to use a newer `numpy` version, which allows it to select a
newer `chex`. But it's easy to imagine similar issues in future with
other packages.

Closes #448.