Build a container with Pallas enabled (#423)
Build a new container that includes dependencies for JAX's
[Pallas](https://jax.readthedocs.io/en/latest/pallas/quickstart.html)
extension: `jax-triton` and `triton`.

Like Pax, T5X, etc. this builds on top of the `jax` container's
"mealkit" step.

`jax-triton` is built from the head of
https://github.com/jax-ml/jax-triton; the commit is recorded in
`manifest.yaml` just like `jax`, and so on.

`triton` is a little more complicated. The upstream repository is
https://github.com/openai/triton, which has a non-default `llvm-head`
branch that is approximately what is used by JAX/OpenXLA, and there is
an OpenXLA fork https://github.com/openxla/triton.

OpenXLA pins ([via Bazel
config](https://github.com/openxla/xla/blob/ea2ad4e91ce28ecfa5c4b8655c8b9b923961fed2/third_party/triton/workspace.bzl#L8))
a tag that exists in this OpenXLA fork; the tag is a snapshot of
`llvm-head` with some extra patches on top. Unfortunately these patches
include the removal of the Python components that are needed by
`jax-triton` and Pallas. The logic in this PR is to roll back the extra
patches from the tag that is pinned in OpenXLA. This yields a version
that does have Python components and should be "close" to the version
hidden inside OpenXLA. This logic is applied when the manifest is
updated, and the resulting commit for `openxla-triton` is saved to
`manifest.yaml` and used when building the Pallas wheel.

Although building the `triton` wheel downloads several GiB of LLVM, the
actual wheel size is only O(150MiB), so the final container is not
dramatically bigger. ARM is not supported.