T5X arm64: Pin grain 0.0.4 source (#520)
Grain's build script recently changed, breaking the way we currently
call said script from the T5X arm64 build. I didn't find a quick way to
fix this on the side of our Dockerfile (short of changing the base
image). Instead, in the interest of unbreaking the T5X build, I pinned
the grain sha corresponding to a recent, published version of the
package, which still works. I'd suggest that going forward we focus on
getting an arm64 wheel of grain published.