Unify JAX and Pallas unit test workflows to overcome the 20-workflow limit in CI (#524)
This is currently done in a sort of 'brute-force' way by stitching
together the two test files and use a workflow input to conditionally
run one of the tests. It is meant to be a quick solution to enable
Pallas test in the new unified workflow.

A more systematic approach can be done in the future.