Update badge links in README (#538)
This is a preliminary update to the landing page badges, due to our
recent transition to a unified nightly/presubmit workflow. Going
forward, we will no longer use any badges generated by GitHub. Instead,
all badge endpoint JSON files are created in the CI workflow on a
per-job basis, and collectively uploaded as GitHub gist by
`_finalize.yaml`.

Known issues to be addressed in follow-up work:
- docker container badges current are not linked to the latest ghcr.io
tags. This is because GitHub container repository uses a 'version
number' to key the tag and that prevents us from having a static link
that point to the latest tag of an image.
- a few badge are missing:
- TE multi-GPU badges are created but having the wrong suffix and hence
not properly recognized during workflow finalization.
- Levanter tests was previously disabled and just enabled by this PR.
The associated badges will be created during the next scheduled run.
  - Rosetta T5X test seems to be failing.