Fix badge publication in CI (#549)
There is a typo that causes deletion of old badges to fail. This was not
captured in the presubmit because the badge publication job did not run.
I've changed it to run regardless whether the workflow is presubmit or
nightly, but let the publishing destination be a mock one if the
workflow is run as presubmit.