Automatic all green bump (#566)
This change introduces a mechanism to automatically bump the manifest &
patches in the triggering branch. So for nightlies that trigger on
`main`, this will bump if all tests have succeeded.

There is now a new input for the unified workflow:
`MERGE_BUMPED_MANIFEST`, which controls whether or not to merge the
manifest bump.
* In scheduled pipelines (nightlies), `MERGE_BUMPED_MANIFEST` is forced
to be `true` and will take a bumped manifest and locally git merge. If
the merge fails or test fails a Draft PR is opened (important in case
there's a merge conflict that requires manual intervention), and then it
deletes the manifest branch to avoid cluttering branch list.
* If `BUMP_MANIFEST=false`, `MERGE_BUMPED_MANIFEST=true`, there will be
an error.

Our two common uses cases are presubmits and nightlies.
* For nightlies, BUMP_MANIFEST/MERGE_BUMPED_MANIFEST are forced to be
`true`.
* For presubmit, the default is taken, which is `false` for both.

Example runs:
| BUMP_MANIFEST | MERGE_BUMPED_MANIFEST | run |
| --- | --- | ---|
| Yes|Yes|https://github.com/NVIDIA/JAX-Toolbox/actions/runs/8023114381|
| Yes|Yes|https://github.com/NVIDIA/JAX-Toolbox/actions/runs/8016123048
(simulate merge conflict - created #584 ) |
|Yes|No|https://github.com/NVIDIA/JAX-Toolbox/actions/runs/8016098571|
|No|No|https://github.com/NVIDIA/JAX-Toolbox/actions/runs/8016080261|
|No|Yes| https://github.com/NVIDIA/JAX-Toolbox/actions/runs/8016095746
(errors)|

And here's an example run where I simulate failure but
`BUMP_MANIFEST`/`MERGE_BUMPED_MANIFEST` are both true:
https://github.com/NVIDIA/JAX-Toolbox/actions/runs/8015902165. In this
case, the [PR is still
made](https://github.com/NVIDIA/JAX-Toolbox/pull/583) in case the error
is deemed acceptable, but needs to be manually merged. FYI, it gets
closed automatically due to the source branch getting deleted, but it
should still be searchable. This helps avoid cluttering our open PRs
view.

## Changes to current workflow
Prior to this PR, the manifest was bumped separately in amd64 and arm64
jobs. This could produce potentially two separate versions of the
manifest, as well as make it difficult to decide which manifest to merge
back in.

The solution taken is to create an artifact with both patches and
manifest and pass that artifact name to the base build.

## Notes about the manifest branch
The manifest branch is always deleted to avoid cluttering the git
branches. Information should not be lost since a PR will be open for
nightlies so that the change-set on the manifest can be recovered and
manually merged.

---------

Co-authored-by: JAX-Toolbox CI <jax@nvidia.com>
Co-authored-by: Yu-Hang "Maxin" Tang <Tang.Maxin@gmail.com>