Add FMHA T5x test (#442)
Adding the JAX T5x FMHA E2E system test to check for fmha lowering
support. Following are the steps implemented in the test:

FMHA lowering flag is enabled by default now, enabled the dumping of hlo
to track fmha forward and backward instructions.
Added the test as part of _ci.yaml file and also added a nightly
workflow file for it. We will add this test as part of performance
benchmarking later and add hlo to baseline.
Also added changes for correction of seq length of decoder (should be a
multiple of 64)

The test was failing with following error related to
CUDNN_STATUS_BAD_PARAM. The fix for this is added in the [PR]
(https://github.com/openxla/xla/pull/6872) in upstream which is now
merged and the test passes.
[Bug](https://nvbugspro.nvidia.com/bug/4409713) for this error.

run for these changes: [workflow run
link](https://github.com/NVIDIA/JAX-Toolbox/actions/runs/7894631992)

---------

Co-authored-by: Terry Kong <terryk@nvidia.com>