nsys-jax: add profiling wrapper script.
This wraps `nsys profile` and writes a .zip archive containing
additional information for offline analysis of JAX/XLA profiles:
- All Python source files referred to in the JAX/XLA metadata.
- XLA metadata exported in .pb and .pbtxt format.
- .proto schema files needed to read the above.

In addition, the .nsys-rep file is compressed inside the .zip archive,
a .parquet format export of the NVTX trace projected onto the GPU
timeline, and a compressed .csv format export of the NVTX trace on the
CPU are included.

The wrapper also ensures that the metadata does not truncate Python
stack traces, and enables CUDA graph node tracing -- although at present
disabling CUDA graphs is recommended to avoid limitations in metadata
correlation.

A Jupyter notebook is bundled in the .zip archive, which provides some
basic analysis out of the box.