ARM64 support workaround (#645)
The arm64 version of the CUDA base container used to build the JAX
containers has an incomplete value of the NVIDIA_REQUIRE_CUDA
environment variable. This prevents the containers from starting on some
systems with older drivers that are, in fact, covered by compatibility
guarantees. This PR copies the NVIDIA_REQUIRE_CUDA value from the
amd64 base container into the arm64 container.


Co-authored-by: Olli Lupton <olupton@nvidia.com>