Disabled gcp autoconfig. (#778)
The `host metadata.google.internal` was causing slow startups on certain
environments. In stead of adding a timeout or changing the check, for
now I recommend we just disable the auto config. There are some other
issues with the current gcp auto config (for instance, incorrectly
relying on $SLURM_JOBID).

I'll take a pass at making that better later this week, but for now lets
turn it off.