nsys-jax: argparse+docs, fix post-processing, analyse collectives (#863)
This improves the command-line error messages and `--help`. The change
to futures handling fixes some non-deterministic behaviour that
sometimes stopped source files from being included in the output
archive.

It also adds new support for calculating message sizes and bandwidths
for collective communication kernels, with an example of how this can be
used in the analysis notebook.