nsys-jax: don't assume recipe produces .sqlite
Previously there were some not-entirely-explicit dependencies between
different parts of post-processing that all used the SQLite export. With
2024.5 this broke, because nsys recipe nvtx_gpu_proj_trace does not produce
an .sqlite export as it used to.

The output data format has changed, and a patch is required to ensure
that it includes the TID with 2024.5..

Add a JAX-based communication test and use it to add end-to-end testing
of nsys-jax in the pipeline. Add a new communication analysis recipe to
allow prompt summarisation of communication data in profiles.

Scale diagram heights better when there are many modules.