nsys-jax: improve alignment for multi-GPU runs
This removes assumptions about the ordering of thunks being identical
across different GPUs, which are not correct in the presence of
overlapped communications and enough jitter.

The summary analysis script gains a wait time metric and prettier formatting.

Also fix a race condition in the EKS-based CI job for nsys-jax, and make
CI unit test jobs go red if some tests fail.