Ensure JAX installation is editable (#1416)
Currently, the JAX Python source that we **installed** in the container
is created by `build-jax.sh` (via _python build.py build --editable ...
--wheels=**jax**,jaxlib,jax-cuda-plugin,jax-cuda-pjrt_) and placed under
`/opt/jaxlibs/jax`:

```
$ docker run --gpus all ghcr.io/nvidia/jax:jax python -c 'import jax; print(jax.__path__)'
['/opt/jaxlibs/jax/jax']
```

As a result, two copies of the JAX source exist in the container at
`/opt/jax` and `/opt/jaxlibs/jax`, leading to potential confusions.
Also, this defeats the 'editable' purpose of the build, as mounting a
JAX working copy from outside the container to `/opt/jax` will not
replace the installed JAX package.

Given that the editable wheel at `/opt/jaxlibs/jax` is practically
identical to the original source (sans a number of build-related files),
we should bypass creating this extra copy and directly specify
`/opt/jax` as the installation target in the Dockerfile and build
script. With this change, we will have
```
>>> import jax; print(jax.__path__)
['/opt/jax/jax']
```
which is the desired usage pattern.