Set --//jaxlib/tools:add_pypi_cuda_wheel_deps=false by default (#1465)
With the merge of https://github.com/jax-ml/jax/pull/28861,
`--build_jaxlib=false` indicates that running the tests will depend on
NVIDIA CUDA wheels hermetically. We don't want that, we want the local
CUDA to be picked by the runtime as we normally do.

https://github.com/jax-ml/jax/pull/28870 provided an flag to turn off
the default setting, we need to have this set for our test script.