triage tool: manage test case output files (#1487)
The main feature added here is that a special directory,
`/triage-tool-output`, is mounted into the container when the test case
is executed, and the host-side versions of this directory are tracked by
the tool so it can identify those corresponding to the last-known-good
and first-known-bad invocations of the test case.

This means that the user can immediately examine "what changed", without
having to manually re-run those two configurations.

Other minor changes:
- Add a `Container` interface with `PyxisContainer` and
`DockerContainer` implementations, for neater type annotations.
- Add a result cache, which allows the tool to avoid some redundant
re-executions
- Write the same log information to `info.log` as to the console
- Only call `git fetch` if needed; if commit-level triage is using a
"new"/failing container, all relevant git history should already be
available.