triage tool: installXXX.sh-based triage (#1505)
This extends support to basic triage of other versioned components
inside the container via external installation scripts. For example, if
the cuBLAS version (as encoded by the `CUBLAS_VERSION` environment
variable) changes from the last-known-good to the first-known-bad
container, the tool can include those two versions in the triage
algorithm and call an external `installCUBLAS.sh`
(`--build-scripts-path`) script to change which one is active.

`--build-scripts-path` is a container *inside* the test
environment/container, so it will likely be combined with
`--container-mount` to allow the installation scripts to be maintained
outside.

Also print a comparison of the environment variables in the
last-known-good and first-known-bad containers if those are both
available.