Triage tool deals with non-linear history (#1538)
If commits do not lie on the same linear git history. The tool would use
`git fetch` because it couldn't find a direct path between the passing
and failing commits.
Here the changes introduce a new logic to find the correct bisection
range:
- apply `git merge-base` to identify common ancestors between the
passing and failing commit
- at build-and-run stage we're checking out the base commit from the
main branch, and we apply cherry-pick for all new relevant features

---------

Co-authored-by: Olli Lupton <olupton@nvidia.com>