Pin orbax-checkpoint to 0.11.19 and pip-tools to 7.4.1 (#1594)
They just so happened to get upgraded on July 31st together, but
- orbax-checkpoint 0.11.20 has issues without internal checkpoint
testing
- pip-tools 7.5.0 will cause `ValueError: '/opt/maxtext/requirements.txt
(line 1)' is not in the subpath of '/opt/pip-tools.d'`. I'm guessing
something is not quite compatible with the Python 3.12 we current have
in the base container. Theoretically, `-r ../maxtext/requirements.txt`
should work, but since we are using a specific version of pip. Let's
play safe at this point and use 7.4.1