build-jax: skip requirements.update when possible (#1732)
The requirements update can be quite slow, and in the normal case it
only serves to propagate the `sed` modifications of
`build/requirements.in` to the requirements lock file. For repeated
re-builds, this is usually a (slow) no-op.

Also allow `CUDA_MAJOR_VERSION` to be set explicitly from the
environment.